Shorten logging keys/values
- Use random int for seed and experiment state ID, instead of long
- Use 'sid' instead of 'stateId' for key
- Remove 'graphType' from MDC

+++ src/main/java/io/sharetrace/experiment/state/BaseExperimentContext.java
@@ -48,7 +48,7 @@ abstract class BaseExperimentContext implements AbstractExperimentContext {
   @Override
   @Value.Default
   public long seed() {
-    return ThreadLocalRandom.current().nextLong(Long.MAX_VALUE);
+    return ThreadLocalRandom.current().nextInt(Integer.MAX_VALUE);
   }
 
   @Override
